Move deserialize tests to primitives
In bitcoin, there are various tests that only have a dependency on
bitcoin due to the deserialize/serialize functions on the old encoding
traits. Changing these to use consensus_encoding allows them to be
trivially moved to primitives.
