+++ bitcoin/src/blockdata/transaction.rs
@@ -1305,10 +1305,9 @@ mod tests {
     use hex_unstable::hex;
 
     use super::*;
-    use crate::consensus::encode::{deserialize, serialize};
+    use crate::consensus::encode::deserialize;
     use crate::constants::WITNESS_SCALE_FACTOR;
     use crate::hex;
-    use crate::script::ScriptSigBuf;
 
     const SOME_TX: &str = "0100000001a15d57094aa7a21a28cb20b59aab8fc7d1149a3bdbcddba9c622e4f5f6a99ece010000006c493046022100f93bb0e7d8db7bd46e40132d1f8242026e045f03a0efe71bbb8e3f475e970d790221009337cd7f1f929f00cc6ff01f03729b069a7c21b59b1736ddfee5db5946c5da8c0121033b9b137ee87d5a812d6f506efdd37f0affa7ffc310711c06c7f3e097c9447c52ffffffff0100e1f505000000001976a9140389035a9225b3839e2bbf32d826a1e222031fd888ac00000000";
 
@@ -1321,12 +1320,6 @@ mod tests {
         let size = tx.consensus_encode(&mut &mut buf[..]).unwrap();
         assert_eq!(size, SOME_TX.len() / 2);
         assert_eq!(raw_tx, &buf[..size]);
-    }
-
-    #[test]
-    fn txin() {
-        let txin: Result<TxIn, _> = deserialize(&hex!("a15d57094aa7a21a28cb20b59aab8fc7d1149a3bdbcddba9c622e4f5f6a99ece010000006c493046022100f93bb0e7d8db7bd46e40132d1f8242026e045f03a0efe71bbb8e3f475e970d790221009337cd7f1f929f00cc6ff01f03729b069a7c21b59b1736ddfee5db5946c5da8c0121033b9b137ee87d5a812d6f506efdd37f0affa7ffc310711c06c7f3e097c9447c52ffffffff"));
-        assert!(txin.is_ok());
     }
 
     #[test]
@@ -1373,14 +1366,6 @@ mod tests {
         assert_eq!(realtx.total_size(), tx_bytes.len());
         assert_eq!(realtx.vsize(), tx_bytes.len());
         assert_eq!(realtx.base_size(), tx_bytes.len());
-    }
-
-    #[test]
-    fn segwit_invalid_transaction() {
-        let tx_bytes = hex!("0000fd000001021921212121212121212121f8b372b0239cc1dff600000000004f4f4f4f4f4f4f4f000000000000000000000000000000333732343133380d000000000000000000000000000000ff000000000009000dff000000000000000800000000000000000d");
-        let tx: Result<Transaction, _> = deserialize(&tx_bytes);
-        assert!(tx.is_err());
-        assert!(matches!(tx.unwrap_err(), crate::consensus::DeserializeError::Parse(_)));
     }
 
     #[test]
@@ -1449,47 +1434,6 @@ mod tests {
         let mut serializer = serde_json::Serializer::new(&mut bytes);
         con_serde::With::<con_serde::Hex>::serialize(&tx, &mut serializer).unwrap();
         assert_eq!(bytes, json.as_bytes())
-    }
-
-    #[test]
-    fn transaction_version() {
-        let tx_bytes = hex!("ffffffff0100000000000000000000000000000000000000000000000000000000000000000000000000ffffffff0100f2052a01000000434104678afdb0fe5548271967f1a67130b7105cd6a828e03909a67962e0ea1f61deb649f6bc3f4cef38c4f35504e51ec112de5c384df7ba0b8d578a4c702b6bf11d5fac00000000");
-        let tx: Result<Transaction, _> = deserialize(&tx_bytes);
-        assert!(tx.is_ok());
-        let realtx = tx.unwrap();
-        assert_eq!(realtx.version, Version::maybe_non_standard(u32::MAX));
-    }
-
-    #[test]
-    fn tx_no_input_deserialization() {
-        let tx_bytes = hex!(
-            "010000000001000100e1f505000000001976a9140389035a9225b3839e2bbf32d826a1e222031fd888ac00000000"
-        );
-        let tx: Transaction = deserialize(&tx_bytes).expect("deserialize tx");
-
-        assert_eq!(tx.inputs.len(), 0);
-        assert_eq!(tx.outputs.len(), 1);
-
-        let reser = serialize(&tx);
-        assert_eq!(tx_bytes, *reser);
-    }
-
-    #[test]
-    fn ntxid() {
-        let tx_bytes = hex!("0100000001a15d57094aa7a21a28cb20b59aab8fc7d1149a3bdbcddba9c622e4f5f6a99ece010000006c493046022100f93bb0e7d8db7bd46e40132d1f8242026e045f03a0efe71bbb8e3f475e970d790221009337cd7f1f929f00cc6ff01f03729b069a7c21b59b1736ddfee5db5946c5da8c0121033b9b137ee87d5a812d6f506efdd37f0affa7ffc310711c06c7f3e097c9447c52ffffffff0100e1f505000000001976a9140389035a9225b3839e2bbf32d826a1e222031fd888ac00000000");
-        let mut tx: Transaction = deserialize(&tx_bytes).unwrap();
-
-        let old_ntxid = tx.compute_ntxid();
-        assert_eq!(
-            format!("{:x}", old_ntxid),
-            "c3573dbea28ce24425c59a189391937e00d255150fa973d59d61caf3a06b601d"
-        );
-        // changing sigs does not affect it
-        tx.inputs[0].script_sig = ScriptSigBuf::new();
-        assert_eq!(old_ntxid, tx.compute_ntxid());
-        // changing pks does
-        tx.outputs[0].script_pubkey = ScriptPubKeyBuf::new();
-        assert!(old_ntxid != tx.compute_ntxid());
     }
 
     #[test]
