+++ primitives/src/block.rs
@@ -1012,6 +1012,8 @@ mod tests {
     #[cfg(feature = "alloc")]
     use encoding::Decode as _;
     use encoding::{Decoder as _, Encode as _, Encoder as _};
+    #[cfg(feature = "hex")]
+    use hex_unstable::hex;
     #[cfg(feature = "alloc")]
     #[cfg(feature = "hex")]
     #[cfg(feature = "serde")]
@@ -2021,5 +2023,41 @@ mod tests {
 
         let roundtrip: Adt = bincode::deserialize(&bytes).expect("failed to deserialize");
         assert_eq!(roundtrip, orig);
+    }
+
+    #[test]
+    #[cfg(feature = "alloc")]
+    #[cfg(feature = "hex")]
+    fn block_version() {
+        let block = hex!("ffffff7f0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000");
+        let decode: Result<Block<Unchecked>, _> = encoding::decode_from_slice(&block);
+        assert!(decode.is_ok());
+
+        let real_decode = decode.unwrap().assume_checked(None);
+        assert_eq!(real_decode.header().version, Version::from_consensus(2147483647));
+
+        let block2 = hex!("000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000");
+        let decode2: Result<Block<Unchecked>, _> = encoding::decode_from_slice(&block2);
+        assert!(decode2.is_ok());
+        let real_decode2 = decode2.unwrap().assume_checked(None);
+        assert_eq!(real_decode2.header().version, Version::from_consensus(-2147483648));
+    }
+
+    #[test]
+    #[cfg(feature = "alloc")]
+    fn static_vector() {
+        // testnet block 000000000000045e0b1660b6445b5e5c5ab63c9a4f956be7e1e69be04fa4497b
+        let segwit_block = include_bytes!("../tests/data/testnet_block_000000000000045e0b1660b6445b5e5c5ab63c9a4f956be7e1e69be04fa4497b.raw");
+        let block: Block<Unchecked> =
+            encoding::decode_from_slice(&segwit_block[..]).expect("failed to deserialize block");
+        assert!(block.check_merkle_root());
+
+        let (header, transactions) = block.into_parts();
+        let block = Block::new_unchecked(header, transactions).assume_checked(None);
+
+        // Same as `block.check_merkle_root` but do it explicitly.
+        let hashes_iter = block.transactions().iter().map(|obj| obj.compute_txid());
+        let from_iter = TxMerkleNode::calculate_root(hashes_iter.clone());
+        assert_eq!(from_iter, Some(block.header().merkle_root));
     }
 }
